+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Please add your own contribution below inside the Master section
 ## Master
 * brief change description - @author
 * expose test file parse result to TestExplorer so the status indicators and test item menu are accessible for all run modes. - @connectdotz
+* added `jest.nodeEnv` and `jest.shell` settings to override default jest process env/shell - @connectdotz
 -->
 
 ### 4.1.2